pi_publisher/requirements.txt
paho-mqtt>=1.6.1
PyYAML>=6.0
sense-hat>=2.6.0